requirements.txt
matplotlib==3.4.2
python-telegram-bot==13.7
requests==2.7.0